src/requirements.txt
paho-mqtt
pyyeelight==1.0-beta
pyyaml
pycrypto